Refactor Task polling (#14863)
## Describe your changes
Currently the way we fetch tasks to be executed is by finding a selection of N tasks, optionally filtering on kind.

For example, when we poll webhook tasks, we just find the latest 10 to run and execute those. It's possible that we've had 10 new tasks get written and the polling from process A doesn't include the task created in process A. This isn't a big deal because every process is polling for the latest 10 so we'll eventually pick up all the tasks

For the adhoc vendor call workflow, it's nicer conceptually (IMO) to write a task, and then execute it in the same http route invocation since we have a dashboard user that is sitting waiting for results.

We will reuse the stuck incode task since this includes executing a workflow to completion

More info on why we're using tasks 2 PRs above

## Caveats

## Test Plan
CI