Represent KBA as list of proven DataLifetimes (#14931)
## Describe your changes
We have this little-used feature for "kba" knowledge-based answers... If you can't log into your account using the verified phone number on file, you can complete some "KBA" in order to instead log in using the unverified email address on your account.

We used to just save the list of `DataIdentifier`s that were proven to be known. Now, we're a bit more explicit and save the `DataLifetime`s. This adds a bit more safety after a future PR where an auth session's user id can be replaced

## Caveats

## Test Plan